Add files via upload
The objective of this project is to produce a model which simulates an operation with real data within Simul8, whilst utilising its range of functions.  Once achieved, we will conduct analysis on the results we observe, and then look to alter components of the system to see if there are practical alterations to the operation that may led to a more efficient outcome. We will then produce a summary of our findings and provide recommendations to the organisation. For us to achieve this effectively, it was imperative that we accessed reliable data, and therefore we aimed to ask relevant and informative questions as part of our survey.  It was also essential that we followed correct procedure when producing the model to ensure its validity and accuracy.
This report can be split into four sections, ‘Research Objectives’, outlining our aims for the model and research interests. ‘Data Collection Methods’, which outlines our approach to collecting the data used for the model, ‘Model Description, Features and Data’ which will discuss the findings from our simulation, explain and visualise the data we have gathered. And, lastly, ‘Simulation Experiments and Results’ which will operate alternative simulations aligned to our research objectives to suggest alterations that can promote efficiency.